refactored scheduler methods to avoid redundancy
